# storage/api-docs/api-docs-admin.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Admin application",
        "description": "API endpoints for the admin application.",
        "contact": {
            "name": "viracoding",
            "email": "viracodingGplus@gmail.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/admin",
            "description": "development Server"
        }
    ],
    "paths": {
...
        "/sessioncourses": {
            "get": {
                "tags": [
                    "SessionCourse"
                ],
                "summary": "Get sessions of courses with pagination",
                "operationId": "9e9074e1c4c21644c0cbe3265e8ed6b1",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SessionCourse"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SessionCourse"
                ],
                "summary": "Store a new session course",
                "operationId": "0fd56e5d54c9d702e70daa730db1e537",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_section_id",
                                    "title",
                                    "description",
                                    "duration_minutes"
                                ],
                                "properties": {
                                    "course_section_id": {
                                        "description": "The ID of the section of course",
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1
                                    },
                                    "title": {
                                        "description": "The title of the session",
                                        "type": "string",
                                        "example": "Session Title"
                                    },
                                    "description": {
                                        "description": "The description of the session",
                                        "type": "string",
                                        "example": "Session Description"
                                    },
                                    "file_url": {
                                        "description": "The file url of the session(externall link like youtube or aparat video link)",
                                        "type": "string",
                                        "example": "http://example.com"
                                    },
                                    "duration_minutes": {
                                        "description": "The duration of the associated course",
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session course created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SessionCourse"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Session course created successfully"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/components/schemas/ValidationError"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/sessioncourses/{id}/file": {
            "post": {
                "tags": [
                    "SessionCourse"
                ],
                "summary": "Update or Insert an existing SessionCourse file",
                "operationId": "481e1d08c32541f0f382f280c6be21da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the session",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "با موفقیت بروزرسانی شد."
                                        },
                                        "file": {
                                            "type": "string",
                                            "example": "http://example.com/storage/sessions"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or image not provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "فایلی ارسال نشده است"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "description": "URL to the first page",
                        "type": "string"
                    },
                    "last": {
                        "description": "URL to the last page",
                        "type": "string"
                    },
                    "prev": {
                        "description": "URL to the previous page",
                        "type": "string"
                    },
                    "next": {
                        "description": "URL to the next page",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "description": "Current page number",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Index of the first item in the current page",
                        "type": "integer"
                    },
                    "last_page": {
                        "description": "Total number of pages",
                        "type": "integer"
                    },
                    "path": {
                        "description": "URL path for the current page",
                        "type": "string"
                    },
                    "per_page": {
                        "description": "Number of items per page",
                        "type": "integer"
                    },
                    "to": {
                        "description": "Index of the last item in the current page",
                        "type": "integer"
                    },
                    "total": {
                        "description": "Total number of items",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "ValidationError",
                "description": "خطای اعتبارسنجی",
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "status": {
                        "type": "integer",
                        "example": 404
                    },
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "دسته والد وجود ندارد!"
                    },
                    "errors": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
...
            "SessionCourse": {
                "title": "SessionCourse",
                "description": "Schema for a session of a course",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the session",
                        "type": "integer",
                        "format": "int64"
                    },
                    "course_section_id": {
                        "description": "The ID of the course section",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "The title of the session",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the session",
                        "type": "string"
                    },
                    "file_path": {
                        "description": "The file path of the session file",
                        "type": "string"
                    },
                    "file_name": {
                        "description": "The file name of the session file",
                        "type": "string"
                    },
                    "file_type": {
                        "description": "The file type of the session file",
                        "type": "string"
                    },
                    "file_size": {
                        "description": "The file size of the session file in bytes",
                        "type": "integer",
                        "format": "int64"
                    },
                    "duration_minutes": {
                        "description": "The duration of the session in minutes",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "The datetime when the session was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The datetime when the session was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}